+++ src/main/java/seed/SeedUtil.java
@@ -188,7 +188,7 @@ public class SeedUtil {
 		int jalpha, ja, jb, jc, jd, je, year, month, day;
 		// double julian = (double) injulian + HALFSECOND / 86400.0;
 
-		ja = (int) injulian;
+		ja = injulian;
 		if (ja >= JGREG) {
 			jalpha = (int) (((ja - 1867216) - 0.25) / 36524.25);
 			ja = ja + 1 + jalpha - jalpha / 4;
